requirements.txt
-f https://download.pytorch.org/whl/cu111/torch_stable.html
torch==1.10.2+cu111
torchvision==0.11.3+cu111
pyyaml==6.0.1
click==8.1.6
numpy==1.24.4
scipy
pandas
matplotlib
tqdm
tensorboard
scikit-learn
Pillow==9.5.0
h5py
opencv-python
imageio==2.31.5
imageio-ffmpeg==0.4.9
streamlit